# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: PhuQuocHub API
  version: 1.0.0
  description: >-
    REST API cho nền tảng tri thức Phú Quốc (Wikipedia + Reddit + Google Maps).
    Đa kênh: Web/Mobile (/api), Public (/public/v1, chỉ đọc), Partner (/partner/v1).
    Mọi phản hồi bọc envelope { success, data, meta } hoặc { success, error }.
  license:
    name: Proprietary
  contact:
    name: PhuQuocHub
servers:
  - url: https://api.phuquochub.example/api
    description: Web/Mobile (nội bộ) — base path /api; versioning /v1 hoãn tới khi ADR-010 (Proposed) được Accepted
  - url: https://api.phuquochub.example/public/v1
    description: Public API (chỉ đọc, API key, ghi công ODbL)
  - url: https://api.phuquochub.example/partner/v1
    description: Partner API (OAuth2 client-credentials)
...
security:
  - bearerAuth: []
paths:
...
  /places/{slug}:
    get:
      tags: [Places]
      summary: Chi tiết địa điểm theo slug
      operationId: getPlace
      security: []
      parameters: [{ $ref: '#/components/parameters/SlugPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties: { data: { $ref: '#/components/schemas/Place' } }
        '404': { $ref: '#/components/responses/NotFound' }
  /places/{id}:
    patch:
      tags: [Places]
      summary: Sửa địa điểm
      operationId: updatePlace
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/PlaceInput' } } }
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Places]
      summary: Lưu trữ địa điểm (soft delete)
      operationId: archivePlace
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses: { '200': { $ref: '#/components/responses/EmptySuccess' } }
...
  /media/presign:
    post:
      tags: [Media]
      summary: Xin presigned PUT URL upload trực tiếp lên object storage (MinIO dev / S3-compatible)
      operationId: presignMedia
      security: [{ bearerAuth: [] }]
      description: >-
        Luồng MỒ CÔI — KHÔNG gắn chủ sở hữu nào (ảnh review, gắn + auto-publish khi tạo review).
        KHÔNG trường owner nào được chấp nhận ở đây (`place_id`/`business_id`/`post_id`/`event_id`/
        `review_id`) — whitelist + forbidNonWhitelisted (main.ts) từ chối 400 nếu gửi.
        `place_id` TỪNG được chấp nhận và đã bị GỠ BỎ (Owner Place Photos, 2026-08-11): nó chỉ
        được kiểm tra "place có tồn tại", KHÔNG kiểm tra người gọi có quyền quản lý cơ sở đó, nên
        bất kỳ member nào cũng gắn được ảnh vào cơ sở của người khác. Ảnh của cơ sở nay đi qua
        `POST /places/{placeId}/media/presign`, nơi place id là route param nên quyền
        `Media.Upload.Managed` được cưỡng chế trên đúng cơ sở đó.
        `key` do SERVER sinh (`media/{uuid-v4}.{ext}`),
        client không tự đặt được. `upload_url` hết hạn sau 600 giây, chỉ dùng được cho `PUT` với
        đúng `Content-Type` đã khai. Đăng ký kết quả qua `POST /media` bên dưới trong vòng phiên
        presign-session (Redis, TTL 900s) — xem docs/data/modules/media.md §Presign-session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type, size, checksum_sha256]
              properties:
                content_type: { type: string, enum: [image/jpeg, image/png, image/webp] }
                size: { type: integer, minimum: 1, maximum: 10485760, description: 'Byte, tối đa 10 MiB.' }
                checksum_sha256:
                  type: string
                  pattern: '^[a-f0-9]{64}$'
                  description: SHA-256 của nội dung file, hex viết thường (64 ký tự) — client tự tính trước khi presign.
      responses:
        '201':
          description: Đã tạo presigned URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          key: { type: string, example: 'media/2f6a1e2e-2b8e-4e9a-9c8b-5b6a7c8d9e0f.jpg' }
                          upload_url: { type: string, format: uri, description: Presigned PUT URL — hết hạn sau expires_in giây, KHÔNG lưu lại/log. }
                          expires_in: { type: integer, example: 600 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /places/{placeId}/media:
    get:
      tags: [Media]
      summary: Ảnh của cơ sở cho màn hình quản lý (MỌI trạng thái)
      operationId: listPlaceMedia
      security: [{ bearerAuth: [] }]
      description: >-
        Khác gallery công khai: trả CẢ `pending`/`rejected`/`hidden` để chủ cơ sở biết ảnh mình gửi
        đang ở đâu trong quy trình. `url` trỏ tới kênh nội bộ `/places/{placeId}/media/{mediaId}/file`
        (ảnh chưa duyệt không có URL công khai nào). KHÔNG trả `object_key`/`bucket`/`checksum`.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/PlaceOwnerMedia' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Media]
      summary: Đăng ký ảnh của cơ sở sau khi PUT (luôn vào pending + hàng chờ kiểm duyệt)
      operationId: registerPlaceMedia
      security: [{ bearerAuth: [] }]
      description: >-
        Cơ sở đích lấy từ PHIÊN PRESIGN (đã khoá `placeId` lúc quyền được kiểm tra), không lấy từ
        body. Tạo dòng media `pending` VÀ một `moderation_cases` (`source=new_content`) trong CÙNG
        một transaction — không bao giờ có ảnh pending mà không ai duyệt.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
                caption: { type: string, maxLength: 300 }
                alt: { type: string, maxLength: 200 }
      responses:
        '201': { description: 'Đã đăng ký — status luôn `pending`, `url` null.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /places/{placeId}/media/presign:
    post:
      tags: [Media]
      summary: Xin presigned PUT URL cho ảnh của cơ sở
      operationId: presignPlaceMedia
      security: [{ bearerAuth: [] }]
      description: >-
        Yêu cầu `Media.Upload.Managed` trên CHÍNH cơ sở ở path (chủ cơ sở hoặc người quản lý được
        gán). Body giống `POST /media/presign` và cũng KHÔNG nhận trường owner nào.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type, size, checksum_sha256]
              properties:
                content_type: { type: string, enum: [image/jpeg, image/png, image/webp] }
                size: { type: integer, minimum: 1, maximum: 10485760 }
                checksum_sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
      responses:
        '201': { description: Đã tạo presigned URL }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /places/{placeId}/media/order:
    patch:
      tags: [Media]
      summary: Sắp xếp lại ảnh của cơ sở
      operationId: reorderPlaceMedia
      security: [{ bearerAuth: [] }]
      description: >-
        Yêu cầu `Media.Upload.Managed` trên CHÍNH cơ sở ở path (chủ cơ sở hoặc người quản lý được
        gán) — cùng đường phân quyền với mọi endpoint ảnh cơ sở khác.


        **Hợp đồng "toàn bộ hay không có gì":** `media_ids` phải là ĐÚNG tập ảnh chưa gỡ của cơ sở
        (MỌI trạng thái, kể cả `pending`/`rejected`), mỗi id đúng MỘT lần. Danh sách thiếu, có id
        trùng, hoặc lẫn id của cơ sở khác đều trả 422 — không có ngữ nghĩa "sắp một phần", vì vị trí
        của ảnh bị bỏ qua sẽ là một câu hỏi không có câu trả lời đúng duy nhất.


        `sort_order` được ghi lại thành 0..n-1 theo đúng thứ tự mảng, trong MỘT câu UPDATE
        set-based, bên trong một transaction có khoá hàng — sắp nhiều lần không làm số trôi dần và
        hai lần sắp đồng thời không lồng vào nhau.


        Sắp ảnh `pending` KHÔNG hề nới lỏng gì ở kênh công khai: gallery công khai vẫn chỉ trả
        `published`; ảnh chờ duyệt chỉ đơn giản đáp xuống đúng vị trí đã chọn khi được duyệt.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [media_ids]
              properties:
                media_ids:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: 'Danh sách ảnh SAU khi sắp, theo thứ tự chuẩn của server.'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/PlaceOwnerMedia' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /places/{placeId}/media/cover:
    patch:
      tags: [Media]
      summary: Chọn ảnh bìa cho cơ sở
      operationId: setPlaceCoverImage
      security: [{ bearerAuth: [] }]
      description: >-
        Yêu cầu `Media.Upload.Managed` trên CHÍNH cơ sở ở path. Client gửi MEDIA ID, KHÔNG BAO GIỜ
        gửi URL: một URL do client cung cấp không kiểm chứng được, còn media id thì kiểm được đầy đủ
        tư cách. Giá trị lưu là `places.cover_image_id`; URL công khai được SINH lúc đọc nên không
        có presigned URL nào bị lưu lại.


        **Điều kiện đủ tư cách** (nằm trong `EXISTS` của chính câu UPDATE, không phải "kiểm tra rồi
        ghi"): ảnh thuộc ĐÚNG cơ sở này, `status = published`, có `object_key`, chưa xoá mềm.


        `404` khi ảnh không thuộc cơ sở này (không phân biệt "không tồn tại" với "của người khác");
        `422` khi ảnh CÓ thuộc cơ sở nhưng chưa được duyệt/đã bị từ chối/đang bị ẩn.


        **Vòng đời:** con trỏ được DỌN tự động khi ảnh bìa bị gỡ (`DELETE .../media/{mediaId}`) hoặc
        bị kiểm duyệt đưa ra khỏi `published` (hide/reject) — cùng transaction với thay đổi gây ra
        nó. KHÔNG tự chọn ảnh thay thế: bìa mới phải do chủ cơ sở chọn tường minh, kể cả sau khi ảnh
        cũ được khôi phục.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [media_id]
              properties:
                media_id: { type: string, format: uuid }
      responses:
        '200':
          description: 'Danh sách ảnh của cơ sở, ảnh bìa mới có `is_cover: true`.'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/PlaceOwnerMedia' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /places/{placeId}/media/{mediaId}:
    delete:
      tags: [Media]
      summary: Gỡ một ảnh khỏi cơ sở (xoá mềm)
      operationId: deletePlaceMedia
      security: [{ bearerAuth: [] }]
      description: >-
        `place_id` nằm trong WHERE của chính câu UPDATE — ảnh thuộc cơ sở khác khớp 0 dòng và trả
        404 (không phân biệt "không tồn tại" với "của người khác").
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: mediaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /places/{placeId}/media/{mediaId}/file:
    get:
      tags: [Media]
      summary: Tải ảnh của cơ sở ở MỌI trạng thái (kênh của chủ cơ sở, redirect tới signed URL)
      operationId: getPlaceMediaFile
      security: [{ bearerAuth: [] }]
      description: >-
        Hai lớp: quyền trên cơ sở (guard) VÀ ảnh phải thuộc đúng cơ sở đó. Trả 302 tới signed URL
        ngắn hạn, `Cache-Control: private`.
      parameters:
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: mediaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '302': { description: Redirect tới signed URL ngắn hạn }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /media/{id}/file:
    get:
      tags: [Media]
      summary: Tải file media đã published (redirect tới signed URL)
      operationId: getMediaFile
      security: []
      description: >-
        Secure Private Media (2026-08-10). Điểm phát media DUY NHẤT ra công chúng — thay cho việc
        trả URL object storage trực tiếp, vốn đòi bucket mở anonymous read cho TOÀN BỘ object (kèm
        cả quyền liệt kê bucket), làm lộ media `pending`/`hidden`/`rejected` cho bất kỳ ai đọc được
        một object key. Bucket nay HOÀN TOÀN riêng tư.


        Công khai có chủ ý (không cần đăng nhập) — khách vãng lai phải xem được ảnh review/gallery.
        Trả `302` tới một presigned GET URL ngắn hạn (mặc định 300s, `S3_PRESIGN_GET_TTL`); KHÔNG
        stream bytes qua API. Trạng thái publish được kiểm tra LẠI ở MỖI request, nên ẩn/từ chối một
        media sẽ thu hồi quyền truy cập ngay ở lần tải kế tiếp.


        `404` cho MỌI trường hợp không phục vụ được — không tồn tại, `pending`, `hidden`,
        `rejected`, đã xoá mềm, hoặc dòng legacy không có `object_key`. Cố ý KHÔNG phân biệt các
        trường hợp này để endpoint công khai không trở thành oracle dò trạng thái kiểm duyệt.


        `Cache-Control: private, max-age=<nửa TTL>` — cho trình duyệt tái dùng redirect trong khi
        signed URL chắc chắn còn hiệu lực, nhưng không cho proxy/CDN dùng chung (signed URL là
        capability dạng bearer).
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '302':
          description: Redirect tới presigned GET URL của object storage.
          headers:
            Location:
              schema: { type: string, format: uri }
              description: Presigned GET URL ngắn hạn (mang chữ ký SigV4).
            Cache-Control:
              schema: { type: string, example: 'private, max-age=150' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /media/{id}:
    delete:
      tags: [Media]
      summary: Xóa media
      operationId: deleteMedia
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses: { '200': { $ref: '#/components/responses/EmptySuccess' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    oauth2Partner:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /partner/v1/oauth/token
          scopes:
            hotels:write: Ghi dữ liệu khách sạn của đối tác
            tours:write: Ghi dữ liệu tour của đối tác
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SlugPath:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
    EmptySuccess:
      description: Thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, const: true }
              data: { type: [object, 'null'] }
...
    BadRequest:
      description: Sai định dạng
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthorized:
      description: Chưa xác thực
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Không đủ quyền
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Không tìm thấy
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: Xung đột (trùng/optimistic lock)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    UnprocessableEntity:
      description: Validation thất bại
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp: { type: string, format: date-time, description: 'ISO timestamp phía server.' }
        requestId: { type: string, description: 'Mã trace/correlation, nếu có.' }
        page: { type: integer, description: 'Trang hiện tại, bắt đầu từ 1.' }
        pageSize: { type: integer, description: 'Số bản ghi mỗi trang SAU khi đã cắt trần (clampLimit).' }
        total: { type: integer, description: 'Tổng số bản ghi khớp bộ lọc.' }
        totalPages: { type: integer, description: 'ceil(total / pageSize); 0 khi pageSize = 0.' }
        quota_remaining:
          type: integer
          deprecated: true
          description: 'KHÔNG được phát ra bởi các endpoint danh sách hiện có.'
        next_cursor:
          type: [string, 'null']
          deprecated: true
          description: >-
            KHÔNG được phát ra — chưa có endpoint nào phân trang bằng cursor. Xem GAP-05/10.
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [PLACE_NOT_FOUND] }
            message: { type: string }
            details: { type: array, items: { type: object, additionalProperties: true } }

    # ---- Enums ----
    PlaceStatus: { type: string, enum: [draft, pending, published, archived] }
    PriceRange: { type: string, enum: [free, low, mid, high] }
    VerificationStatus: { type: string, enum: [pending, verified, official, community_verified, expired, rejected] }
...
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
...
    PlaceCard:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        category_id: { type: string, format: uuid }
        location: { $ref: '#/components/schemas/GeoPoint' }
        short_description: { type: [string, 'null'] }
        price_range: { $ref: '#/components/schemas/PriceRange' }
        # Ảnh bìa do chủ cơ sở chọn (`places.cover_image_id` → media), phân giải Ở TẦNG ĐỌC thành
        # URL API ỔN ĐỊNH `{API_PUBLIC_URL}/{prefix}/media/{id}/file` — KHÔNG BAO GIỜ là presigned
        # URL hay địa chỉ object storage, và không có URL nào được lưu sẵn trong DB. `null` khi cơ
        # sở chưa chọn bìa HOẶC khi ảnh bìa không (còn) đủ điều kiện: đường đọc lọc lại
        # `published` + đúng cơ sở + chưa xoá mềm một cách độc lập với đường ghi, nên ảnh
        # pending/rejected/hidden không bao giờ lộ ra đây. Xem core/media-url/cover-image.ts.
        cover_image_url: { type: [string, 'null'], format: uri }
        rating_avg: { type: [number, 'null'] }
        rating_count: { type: integer }
        verification_status: { $ref: '#/components/schemas/VerificationStatus' }
        # F-17 / OD-F-17 (2026-07-23): runtime LUÔN phát trường này (CARD_COLS chọn p.status;
        # places.mapper.ts đặt vô điều kiện) nhưng schema trước đây không khai. Đây là TRẠNG THÁI
        # XUẤT BẢN, KHÔNG phải giờ mở cửa của cơ sở (giờ mở cửa nằm ở `opening_hours`). Dùng đúng
        # enum PlaceStatus đã có sẵn — enum này vốn ĐÃ công khai qua schema `Place` bên dưới, nên
        # việc khai ở đây không lộ thêm bất kỳ thông tin nào và payload không đổi một byte.
        # Kênh công khai chỉ có thể quan sát `published` vì read công khai lọc theo trạng thái đó.
        status: { $ref: '#/components/schemas/PlaceStatus' }
        distance_m: { type: [number, 'null'] }
    Place:
      allOf:
        - $ref: '#/components/schemas/PlaceCard'
        - type: object
          properties:
            # Slug danh mục — CHỈ có ở chi tiết, không có ở PlaceCard. `category_id` một mình là
            # UUID mà client không dùng được nếu không gọi thêm GET /categories; slug cho phép
            # trang chi tiết điều hướng về đúng trang duyệt (vd 'attraction' → /attractions).
            category_slug: { type: [string, 'null'] }
            address: { type: [string, 'null'] }
            ward: { type: [string, 'null'] }
            description: { type: [string, 'null'] }
            opening_hours: { type: [object, 'null'], additionalProperties: true }
            status: { $ref: '#/components/schemas/PlaceStatus' }
            osm_id: { type: [integer, 'null'] }
            contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } }
            prices: { type: array, items: { $ref: '#/components/schemas/Price' } }
            media: { type: array, items: { $ref: '#/components/schemas/Media' } }
            faqs: { type: array, items: { type: object, additionalProperties: true } }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    PlaceInput:
      type: object
      required: [name, category_id, location]
      properties:
        name: { type: string, maxLength: 200 }
        category_id: { type: string, format: uuid }
        location: { $ref: '#/components/schemas/GeoPoint' }
        address: { type: string }
        ward: { type: string }
        description: { type: string }
        short_description: { type: string, maxLength: 300 }
        opening_hours: { type: object, additionalProperties: true }
        price_range: { $ref: '#/components/schemas/PriceRange' }

    # ---- Transport (ADR-017, Accepted 2026-07-28) ----
    # Place có category='transport' + satellite `place_transport_details` (1:1). Trục phân loại
    # chính là từ điển `transport_types` (GET /transport-types), không phải ENUM. Nhiệm vụ nền
    # tảng: các schema dưới đây phản ánh ĐÚNG những gì `GET /transports`/`{slug}` thực sự trả về
    # — không tài liệu hoá endpoint hay trường nào chưa triển khai.
...
    Contact:
      type: object
      properties:
        id: { type: string, format: uuid }
        owner_type: { type: string, examples: [place, business] }
        contact_type: { type: string, examples: [HOTLINE, PHONE, EMAIL, WEBSITE, FACEBOOK] }
        value: { type: string }
        label: { type: [string, 'null'] }
        is_primary: { type: boolean }
        verification_status: { $ref: '#/components/schemas/VerificationStatus' }
        display_order: { type: integer }
...
    Price:
      type: object
      properties:
        id: { type: string, format: uuid }
        service_name: { type: string }
        amount: { type: number }
        currency: { type: string, examples: [VND] }
        unit: { type: [string, 'null'] }
        is_free: { type: boolean }
        valid_from: { type: [string, 'null'], format: date-time }
        valid_to: { type: [string, 'null'], format: date-time }
        verification_status: { $ref: '#/components/schemas/VerificationStatus' }
...
    PlaceOwnerMedia:
      type: object
      description: >-
        Ảnh của cơ sở như CHỦ CƠ SỞ nhìn thấy (Owner Place Photos). Khác `Media` công khai: có MỌI
        trạng thái, và `url` trỏ tới kênh nội bộ theo cơ sở vì ảnh chưa duyệt không có URL công
        khai. KHÔNG có `object_key`/`bucket`/`checksum_sha256` — chi tiết lưu trữ không rời server.
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, published, rejected, hidden] }
        caption: { type: [string, 'null'] }
        alt_text: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        url:
          type: string
          description: '`/places/{placeId}/media/{id}/file` — kiểm tra quyền lại mỗi lần tải.'
        sort_order:
          type: [integer, 'null']
          description: >-
            Vị trí trong gallery, 0 = đầu tiên (Owner Cover & Photo Ordering). `null` = chưa từng
            được sắp — xếp SAU mọi ảnh đã sắp. Ghi qua `PATCH /places/{placeId}/media/order`.
        is_cover:
          type: boolean
          description: >-
            Ảnh này có đang là `places.cover_image_id` không. Chỉ ĐÚNG MỘT ảnh mỗi cơ sở; suy ra từ
            `places`, không phải cột riêng trên media.
    Media:
      type: object
      description: >-
        `url` là `null` cho media `pending` (mọi upload mới qua Media Upload Foundation, 2026-07-30
        — không có object_key/bucket/... nào được lộ, chỉ metadata storage tối thiểu được LƯU,
        không bao giờ là URL tuyệt đối/signed — xem docs/data/modules/media.md §Persistence). Chỉ
        media `published` hợp pháp (embed YouTube/Vimeo, hoặc một luồng kiểm duyệt tương lai) mới
        có `url` thật.


        Secure Private Media (2026-08-10): với media upload đã `published`, `url` là endpoint API
        ỔN ĐỊNH `{API_PUBLIC_URL}/api/media/{id}/file` — KHÔNG phải URL object storage, KHÔNG mang
        chữ ký, KHÔNG hết hạn. Nó redirect (302) tới một signed URL ngắn hạn ở mỗi lần tải, sau khi
        kiểm tra lại trạng thái publish. `object_key`/`bucket` không bao giờ rời server.
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [image, video] }
        url: { type: [string, 'null'], format: uri }
        thumbnail_url: { type: [string, 'null'], format: uri }
        caption: { type: [string, 'null'] }
        alt_text: { type: [string, 'null'] }
        status: { type: string, enum: [pending, published, hidden, rejected] }
...